day1206

+++ zeno/tpls/include/tinyexr.h
@@ -11264,7 +11264,7 @@ int LoadEXR(float **out_rgba, int *width, int *height, const char *filename,
   }
 
   // Read HALF channel as FLOAT.
-  for (int i = 0; i < exr_header.num_channels; i++) {
+  for (size_t i = 0; i < exr_header.num_channels; i++) {
     if (exr_header.pixel_types[i] == TINYEXR_PIXELTYPE_HALF) {
       exr_header.requested_pixel_types[i] = TINYEXR_PIXELTYPE_FLOAT;
     }
@@ -11283,7 +11283,7 @@ int LoadEXR(float **out_rgba, int *width, int *height, const char *filename,
   int idxG = -1;
   int idxB = -1;
   int idxA = -1;
-  for (int c = 0; c < exr_header.num_channels; c++) {
+  for (size_t c = 0; c < exr_header.num_channels; c++) {
     if (strcmp(exr_header.channels[c].name, "R") == 0) {
       idxR = c;
     } else if (strcmp(exr_header.channels[c].name, "G") == 0) {
@@ -11303,14 +11303,14 @@ int LoadEXR(float **out_rgba, int *width, int *height, const char *filename,
                static_cast<size_t>(exr_image.height)));
 
     if (exr_header.tiled) {
-      for (int it = 0; it < exr_image.num_tiles; it++) {
-        for (int j = 0; j < exr_header.tile_size_y; j++) {
-          for (int i = 0; i < exr_header.tile_size_x; i++) {
-            const int ii =
+      for (size_t it = 0; it < exr_image.num_tiles; it++) {
+        for (size_t j = 0; j < exr_header.tile_size_y; j++) {
+          for (size_t i = 0; i < exr_header.tile_size_x; i++) {
+            const size_t ii =
                 exr_image.tiles[it].offset_x * exr_header.tile_size_x + i;
-            const int jj =
+            const size_t jj =
                 exr_image.tiles[it].offset_y * exr_header.tile_size_y + j;
-            const int idx = ii + jj * exr_image.width;
+            const size_t idx = ii + jj * exr_image.width;
 
             // out of region check.
             if (ii >= exr_image.width) {
@@ -11333,7 +11333,7 @@ int LoadEXR(float **out_rgba, int *width, int *height, const char *filename,
         }
       }
     } else {
-      for (int i = 0; i < exr_image.width * exr_image.height; i++) {
+      for (size_t i = 0; i < exr_image.width * exr_image.height; i++) {
         const float val = reinterpret_cast<float **>(exr_image.images)[0][i];
         (*out_rgba)[4 * i + 0] = val;
         (*out_rgba)[4 * i + 1] = val;
@@ -11370,9 +11370,9 @@ int LoadEXR(float **out_rgba, int *width, int *height, const char *filename,
         malloc(4 * sizeof(float) * static_cast<size_t>(exr_image.width) *
                static_cast<size_t>(exr_image.height)));
     if (exr_header.tiled) {
-      for (int it = 0; it < exr_image.num_tiles; it++) {
-        for (int j = 0; j < exr_header.tile_size_y; j++) {
-          for (int i = 0; i < exr_header.tile_size_x; i++) {
+      for (size_t it = 0; it < exr_image.num_tiles; it++) {
+        for (size_t j = 0; j < exr_header.tile_size_y; j++) {
+          for (size_t i = 0; i < exr_header.tile_size_x; i++) {
             const int ii =
                 exr_image.tiles[it].offset_x * exr_header.tile_size_x + i;
             const int jj =
@@ -11404,7 +11404,7 @@ int LoadEXR(float **out_rgba, int *width, int *height, const char *filename,
         }
       }
     } else {
-      for (int i = 0; i < exr_image.width * exr_image.height; i++) {
+      for (size_t i = 0; i < exr_image.width * exr_image.height; i++) {
         (*out_rgba)[4 * i + 0] =
             reinterpret_cast<float **>(exr_image.images)[idxR][i];
         (*out_rgba)[4 * i + 1] =
